Updated halfix with latest changes from lib86cpu

+++ include/lib86cpu/cpu.h
@@ -6,11 +6,11 @@
 
 
 int cpu_init(struct pc_settings *pc);
-void cpu_destroy_io_region(uint32_t port, uint32_t size);
-lc86_status cpu_add_io_region(port_t port, uint32_t size, io_handlers_t handlers, void *opaque);
-lc86_status cpu_add_mmio_region(addr_t addr, uint32_t size, io_handlers_t handlers, void *opaque);
-lc86_status cpu_add_ram_region(addr_t addr, uint32_t size);
-lc86_status cpu_add_rom_region(addr_t addr, uint32_t size, uint8_t *buffer);
+void cpu_destroy_io_region(uint32_t port, uint64_t size);
+lc86_status cpu_add_io_region(port_t port, uint64_t size, io_handlers_t handlers, void *opaque);
+lc86_status cpu_add_mmio_region(addr_t addr, uint64_t size, io_handlers_t handlers, void *opaque);
+lc86_status cpu_add_ram_region(addr_t addr, uint64_t size);
+lc86_status cpu_add_rom_region(addr_t addr, uint64_t size, uint8_t *buffer);
 void cpu_pause();
 void cpu_resume();
